requirements.txt
websockets
aiohttp
aiokafka
kafka-python
python-dotenv
requests
asyncpg
streamlit
plotly
ta
